Avoid the NonEmptyPromise
The NonEmptyPromise crashes tests with an assert that is hard to understand, especially in CI environments.
The intention is to make sure that an unused URLSessionMock is detected in a test run, but there are better
solutions for this.

With this change, the passthrough subject will emit a "canceled" error when the subscription ends and nothing
was sent through it. This behavior is not any closer to a real URL session, but it does allow the error to be
bubbled up to the consumer of the session, and the test will fail in a clear and understandable way.